Use single-arg lambdas in the AST
This commit updates the AST to treat lambdas as always having a single
argument. This simplifies typechecking (& will do so even more once the
bidirectional typechecker is implemented).

The surface syntax is unaffected, so from a UX perspective it still
appears that we support multi-argument lambdas.

We still avoid the creation of unnecessary nested closures in the
bytecode representation by "squashing" single argument lambdas to
multi-arg ones in the compiler.