Intern variable names
This commit adds an `interner` module, with a single public `Id` type,
which can be used to globally intern strings, and then uses it to intern
variable names during parsing.

Currently nothing further is done with these names, but since they're
just `usize`s they're cheap to copy around so we can keep them in the
AST & bytecode in order to name variables during error reporting.